+++ documentation/ChangeLog.md
@@ -3,6 +3,11 @@
 
 ### (NextVersion)
  * Fixed: Arguments field in launch configuration is not multi-line.
+ * Added support for Build Configurations.
+   * Available build targets are displayed in the Project Explorer. 
+   You can configure which targets are enabled for a workspace build or not. Or run/debug a specific target.
+   * Default build targets are: #TODO
+   * Project Build Configuration property page update to configure build targets.
  
   ▶ Recommended/tested CDT version is now 8.7.0